Added Rigidbody and changed positioning to vectors (Vector3).  Added physics to Rigidbody to allow bird to behave more like the flappy bird game.

+++ GameTutorial/Engine/Engine.cpp
@@ -5,7 +5,7 @@
 int Engine::SCREEN_WIDTH = 1024;
 int Engine::SCREEN_HEIGHT = 768;
 GLFWwindow* Engine::window = NULL;
-double Engine::dt = 0;
+float Engine::dt = 0;
 
 
 Engine::Engine()
@@ -66,14 +66,14 @@ bool Engine::Initialize(const char* windowTitle)
 	glEnable(GL_BLEND);
 	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 
-	lastTime = glfwGetTime();
+	lastTime = (float)glfwGetTime();
 
 	return true;
 }
 
 void Engine::Update()
 {
-	double now = glfwGetTime();
+	float now = (float)glfwGetTime();
 	dt = (now - lastTime);
 	lastTime = now;
 	glfwPollEvents();
@@ -90,7 +90,7 @@ void Engine::EndRender()
 	glfwSwapBuffers(window);
 }
 
-double Engine::GetDT()
+float Engine::GetDT()
 {
 	return dt;
 }
